requirements.txt
Flask
requests
Pillow
piexif
python-dotenv
